- Add redux-thunk - Move store configuration into separate file - Purge unused redux-devtools-* packages
