--- app/client/actions/index.js
@@ -1,5 +1,0 @@
-import user from './user';
-
-export default { 
-  user
-};
