+++ package.json
@@ -12,7 +12,8 @@
     "react-dom": "^15.4.2",
     "react-redux": "^5.0.2",
     "redux": "^3.0.0",
-    "redux-actions": "^1.2.1"
+    "redux-actions": "^1.2.1",
+    "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
     "babel-cli": "^6.22.2",
@@ -32,9 +33,7 @@
     "eslint": "^3.14.1",
     "eslint-config-airbnb": "^14.0.0",
     "eslint-plugin-react": "^6.9.0",
-    "mocha": "^3.2.0",
-    "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-devtools-log-monitor": "^1.2.0"
+    "mocha": "^3.2.0"
   },
   "scripts": {
     "serve": "babel-node scripts/serve.js",
